+++ pkg/providers/clientlists/testData/TestResActivation/activation_update_version_only.tf
@@ -4,7 +4,6 @@ provider "akamai" {
 
 resource "akamai_clientlist_activation" "activation_ASN_LIST_1" {
   list_id                 = "12_AB"
-  version                 = 3
   network                 = "STAGING"
   comments                = "Activation Comments"
   notification_recipients = ["user@example.com"]