+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 #### BREAKING CHANGES:
 
+* ClientLists
+    * Made the following changes in the `resource_akamai_clientlists_list_activation` resource:
+    * Marked `version` attribute in schema as `Computed` instead of `Required`.
 
 * PAPI
   * Removed the `secret_key` and `api_key` attributes from the `akamai_property_rule_formats` data source for the rule formats `v2025-05-30`, `v2025-07-07`, `v2025-09-09` and `v2025-10-16`.
@@ -63,8 +66,6 @@
 * PAPI
     * Added support for the new rule format [`v2026-01-09`](https://techdocs.akamai.com/terraform/docs/rule-format-changes#v2026-01-09).
     * Introduced support for the `api_key_cam_guid`, `additional_headers_mode`, and `additional_headers_list` attributes in the `akamai_property_rule_formats` data source for the rule formats `v2025-05-30`, `v2025-07-07`, `v2025-09-09` and `v2025-10-16`.
-
-